refactor: streamline client connection logic and update configuration structure
