+++ src/messages/events/client/initialize.rs
@@ -1,6 +1,8 @@
 use std::sync::Arc;
 use std::sync::atomic::Ordering;
 
+use crate::error::CordErrorKind;
+use crate::ipc::discord::client::Connection;
 use crate::messages::events::event::{EventContext, OnEvent};
 use crate::types::config::PluginConfig;
 use crate::util::now;
@@ -30,6 +32,38 @@ impl OnEvent for InitializeEvent {
                 Ordering::SeqCst,
                 Ordering::SeqCst,
             );
+        }
+
+        let rich_client = &ctx.cord.rich_client;
+        let mut client = rich_client.write().unwrap();
+        if !self.config.advanced.discord.pipe_paths.is_empty()
+            && client.pipe_paths.is_empty()
+        {
+            client.pipe_paths = self.config.advanced.discord.pipe_paths.clone();
+        }
+
+        let config = &ctx.cord.config;
+        if !client.is_ready.load(Ordering::SeqCst) && client.connect().is_err()
+        {
+            if config.reconnect_interval > 0 && config.initial_reconnect {
+                drop(client);
+                let client_clone = rich_client.clone();
+                let tx = ctx.cord.tx.clone();
+
+                let reconnect_interval = config.reconnect_interval;
+                std::thread::spawn(move || {
+                    let mut client = client_clone.write().unwrap();
+                    client.reconnect(reconnect_interval, tx.clone());
+                });
+            } else {
+                return Err(crate::error::CordError::new(
+                    CordErrorKind::Io,
+                    "Failed to connect to Discord",
+                ));
+            }
+        } else {
+            client.handshake()?;
+            client.start_read_thread(ctx.cord.tx.clone())?;
         }
 
         if let Some(mut session) =
